refactor: group SortableList and SortableListItem into a sortable-list/ folder
