+++ resources/assets/js/components/ui/sortable-list/SortableListItem.vue
